progress bar test, files renamed
