+++ src/components/basic-questions/BasicQuestions.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
 import { Button, Form } from "react-bootstrap";
 import { Basic_Question } from "../../interfaces/basic-question";
-import { ProgressBar } from "../progress-bar/progress-bar";
+import { ProgressBar } from "../progress-bar/progressBar";
 
 import basicData from "../../data/basic-questions.json"
 import "./BasicQuestions.css"
